decoupled the logic to list user activities from the MyActivity page
